+++ MicroserviceApp/Backend/AdRequestManager/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 #MYSQL CREDENTIAL FOR CLASSIC RUN
-spring.datasource.username = agentXML
-spring.datasource.password = agent1
-spring.datasource.url = jdbc:mysql://localhost:3306/agent_db?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
+spring.datasource.username =
+spring.datasource.password = 
+spring.datasource.url = jdbc:mysql://localhost:3306/_db?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
 
 #MYSQL CONFIG
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
